requirements.txt
# Для запуска проекта:
# pip install -r requirements.txt

asgiref==3.8.1
django==4.2.23
djangorestframework==3.16.0
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
